Dev prep 1.2.1 release (#415)
* Remove dependency copying (#376)

There doesn't seem to be a reason to have this phase. Perhaps there was
one before? All the required dependencies are added to the final JAR by
the `maven-shade-plugin` in the `package` phase anyway. This dependency
copying just seems like a waste of time.

* Add .editorconfig

Provides common whitespace settings for all editors. Supported out of
the box in some IDE's (like IntelliJ IDEA) and all others have
appropriate plugins.

*  Added .cflintrc schema and output schemas for JSON and XML. Tweaked README. (#377)

Added .cflintrc schema and output schemas for JSON and XML. Tweaked README.

* support autochange

* #382

* Align Gradle dependencies with Maven ones (#378)

Define the same set (and scopes) of dependencies for Gradle builds as we
have now in Maven's `pom.xml`.

* Added some context to the changelog for 1.2.0 and removed a lot of "##### no issues" items that were added to the list.

* Changed setting for auto-adding a final new line, that screws up expected files for tests.

* #383 Changed code, test, README with the output sample and JSON schema.

* Update timestamp type in result schemas

* #385 Two small modifications to HTML report

* Changed build versions to 1.2.1-SNAPSHOT in dev branch and change documentation to 1.2.1 in dev branch.

* 390 Add Implementation-Version to MANIFEST

Configures the Maven and Gradle builds to include
`Implementation-Version` in the MANIFEST.MF file, which can then be read
by the application and displayed to the user.

* #385 - Further HTML report improvements

* #390 - Adapting new way to grab version string

* #390 - Hopefully fixing Gradle build

* #390 - Review changes

* #396 fixed

* Update README.md

Explains how to pass multiple arguments in the command line in Windows Powershell and fixes a typo.

Fixes #400

* Update CODE_OF_CONDUCT.md

* #380 fixed

* Added VS Code info. Some cleanup.

* use cfparser 2.4.11

* #408

* #373

* #367

* #367

* add sonatype repo

* merged master

* cfparser 2.5.0

* Added test for #359 showing parsing error gone with cfparser 2.5.0

* #359 use parseCFMLExpression

* fix tests

* test for #346

* #413

* fix test

* fix javadoc